Load a selection of EPI CPS extracts

Select years and variables from the EPI CPS microdata extracts. These data must first be downloaded using download_cps() or from https://microdata.epi.org.

load_cps(.sample, .years, ..., .extracts_dir = NULL, .version_check = TRUE)

load_basic(.years, ..., .extracts_dir = NULL, .version_check = TRUE)

load_may(.years, ..., .extracts_dir = NULL, .version_check = TRUE)

load_org(.years, ..., .extracts_dir = NULL, .version_check = TRUE)

load_org_sample(.years, ..., .extracts_dir = NULL, .version_check = TRUE)

Arguments

.sample

CPS sample ("org", "basic", "march", "may")

.years

years of CPS data (integers), or a cps_files object

...

tidy selection of variables to keep

.extracts_dir

directory where EPI extracts are

.version_check

when TRUE, confirm data are same version

Value

A tibble of CPS microdata

Details

All columns are selected if ... is missing.

.years accepts either integer years or the result of cps_files(). When a cps_files object is passed, the files are read directly and .extracts_dir is ignored.

.extracts_dir is required, but if NULL it will look for the environment variables

EPIEXTRACTS_CPSBASIC_DIR
EPIEXTRACTS_CPSORG_DIR
EPIEXTRACTS_CPSMAY_DIR

which could be set in your .Renviron, for example.

Functions

load_cps(): base function group

load_basic(): Load CPS Basic Monthly files

load_may(): Load CPS May files

load_org(): Load CPS ORG files

load_org_sample(): Load a demonstration sample of CPS ORG files; only useful for examples

Examples

# Load all columns from the demonstration sample
load_org_sample(2019:2020)
#> Using Demonstration sample EPI CPS ORG Extracts, Version 1.0.55
#> # A tibble: 560,340 × 11
#>     year month orgwgt statefips wbho      female   educ     wage wageotc emp    
#>    <int> <int>  <dbl> <int+lbl> <int+lbl> <int+lb> <int+l> <dbl>   <dbl> <int+l>
#>  1  2019     1 11367. 1 [AL]    2 [Black] 1 [Fema… 3 [Som… 14      14    1 [Emp…
#>  2  2019     1  6541. 1 [AL]    2 [Black] 1 [Fema… 3 [Som… 20.9    20.9  1 [Emp…
#>  3  2019     1  6327. 1 [AL]    2 [Black] 0 [Male] 1 [Les…  7.65    7.65 1 [Emp…
#>  4  2019     1  6327. 1 [AL]    2 [Black] 0 [Male] 1 [Les…  7.65    7.65 1 [Emp…
#>  5  2019     1 11262. 1 [AL]    2 [Black] 1 [Fema… 3 [Som… 10      10    1 [Emp…
#>  6  2019     1  7867. 1 [AL]    2 [Black] 1 [Fema… 4 [Col… 28.8    28.8  1 [Emp…
#>  7  2019     1 11262. 1 [AL]    2 [Black] 1 [Fema… 3 [Som… 11      12.2  1 [Emp…
#>  8  2019     1  7943. 1 [AL]    2 [Black] 0 [Male] 3 [Som… NA      NA    0 [NIL…
#>  9  2019     1  6092. 1 [AL]    2 [Black] 1 [Fema… 1 [Les… NA      NA    0 [NIL…
#> 10  2019     1  7738. 1 [AL]    1 [White] 0 [Male] 2 [Hig… NA      NA    0 [NIL…
#> # ℹ 560,330 more rows
#> # ℹ 1 more variable: lfstat <int+lbl>

# Load a selection of columns
load_org_sample(2019:2023, year, month, female, wage)
#> Using Demonstration sample EPI CPS ORG Extracts, Version 1.0.55
#> # A tibble: 1,318,922 × 4
#>     year month female      wage
#>    <int> <int> <int+lbl>  <dbl>
#>  1  2019     1 1 [Female] 14   
#>  2  2019     1 1 [Female] 20.9 
#>  3  2019     1 0 [Male]    7.65
#>  4  2019     1 0 [Male]    7.65
#>  5  2019     1 1 [Female] 10   
#>  6  2019     1 1 [Female] 28.8 
#>  7  2019     1 1 [Female] 11   
#>  8  2019     1 0 [Male]   NA   
#>  9  2019     1 1 [Female] NA   
#> 10  2019     1 0 [Male]   NA   
#> # ℹ 1,318,912 more rows

# These are equivalent:
load_org_sample(2019:2020)
#> Using Demonstration sample EPI CPS ORG Extracts, Version 1.0.55
#> # A tibble: 560,340 × 11
#>     year month orgwgt statefips wbho      female   educ     wage wageotc emp    
#>    <int> <int>  <dbl> <int+lbl> <int+lbl> <int+lb> <int+l> <dbl>   <dbl> <int+l>
#>  1  2019     1 11367. 1 [AL]    2 [Black] 1 [Fema… 3 [Som… 14      14    1 [Emp…
#>  2  2019     1  6541. 1 [AL]    2 [Black] 1 [Fema… 3 [Som… 20.9    20.9  1 [Emp…
#>  3  2019     1  6327. 1 [AL]    2 [Black] 0 [Male] 1 [Les…  7.65    7.65 1 [Emp…
#>  4  2019     1  6327. 1 [AL]    2 [Black] 0 [Male] 1 [Les…  7.65    7.65 1 [Emp…
#>  5  2019     1 11262. 1 [AL]    2 [Black] 1 [Fema… 3 [Som… 10      10    1 [Emp…
#>  6  2019     1  7867. 1 [AL]    2 [Black] 1 [Fema… 4 [Col… 28.8    28.8  1 [Emp…
#>  7  2019     1 11262. 1 [AL]    2 [Black] 1 [Fema… 3 [Som… 11      12.2  1 [Emp…
#>  8  2019     1  7943. 1 [AL]    2 [Black] 0 [Male] 3 [Som… NA      NA    0 [NIL…
#>  9  2019     1  6092. 1 [AL]    2 [Black] 1 [Fema… 1 [Les… NA      NA    0 [NIL…
#> 10  2019     1  7738. 1 [AL]    1 [White] 0 [Male] 2 [Hig… NA      NA    0 [NIL…
#> # ℹ 560,330 more rows
#> # ℹ 1 more variable: lfstat <int+lbl>
load_cps("org_sample", 2019:2020)
#> Using Demonstration sample EPI CPS ORG Extracts, Version 1.0.55
#> # A tibble: 560,340 × 11
#>     year month orgwgt statefips wbho      female   educ     wage wageotc emp    
#>    <int> <int>  <dbl> <int+lbl> <int+lbl> <int+lb> <int+l> <dbl>   <dbl> <int+l>
#>  1  2019     1 11367. 1 [AL]    2 [Black] 1 [Fema… 3 [Som… 14      14    1 [Emp…
#>  2  2019     1  6541. 1 [AL]    2 [Black] 1 [Fema… 3 [Som… 20.9    20.9  1 [Emp…
#>  3  2019     1  6327. 1 [AL]    2 [Black] 0 [Male] 1 [Les…  7.65    7.65 1 [Emp…
#>  4  2019     1  6327. 1 [AL]    2 [Black] 0 [Male] 1 [Les…  7.65    7.65 1 [Emp…
#>  5  2019     1 11262. 1 [AL]    2 [Black] 1 [Fema… 3 [Som… 10      10    1 [Emp…
#>  6  2019     1  7867. 1 [AL]    2 [Black] 1 [Fema… 4 [Col… 28.8    28.8  1 [Emp…
#>  7  2019     1 11262. 1 [AL]    2 [Black] 1 [Fema… 3 [Som… 11      12.2  1 [Emp…
#>  8  2019     1  7943. 1 [AL]    2 [Black] 0 [Male] 3 [Som… NA      NA    0 [NIL…
#>  9  2019     1  6092. 1 [AL]    2 [Black] 1 [Fema… 1 [Les… NA      NA    0 [NIL…
#> 10  2019     1  7738. 1 [AL]    1 [White] 0 [Male] 2 [Hig… NA      NA    0 [NIL…
#> # ℹ 560,330 more rows
#> # ℹ 1 more variable: lfstat <int+lbl>

# Use cps_files() for targets workflows:
load_org_sample(cps_files("org_sample", 2019:2023), year, month, wage)
#> Using Demonstration sample EPI CPS ORG Extracts, Version 1.0.55
#> # A tibble: 1,318,922 × 3
#>     year month  wage
#>    <int> <int> <dbl>
#>  1  2019     1 14   
#>  2  2019     1 20.9 
#>  3  2019     1  7.65
#>  4  2019     1  7.65
#>  5  2019     1 10   
#>  6  2019     1 28.8 
#>  7  2019     1 11   
#>  8  2019     1 NA   
#>  9  2019     1 NA   
#> 10  2019     1 NA   
#> # ℹ 1,318,912 more rows

if (FALSE) { # \dontrun{
# Load real CPS ORG data (requires downloaded extracts):
load_org(2010:2019, year, month, orgwgt, female, wage)
} # }